Add CSV in iteration directory for persons on links
